// src/Type/Types.ts
export interface Review {
  rating: number;
  comment: string;
  reviewerName: string;
  reviewerEmail: string;
}

export interface Product {
  id: number;
  title: string;
  description: string;
  category: string;
  price: number;
  brand: string;
  rating: number;
  thumbnail: string;
  reviews: Review[];
}

export interface ProductsResponse {
  products: Product[];
  total: number;
  skip: number;
  limit: number;
}

export interface Category {
  slug: string;
  name: string;
  url: string;
}

export interface RouteParams {
  id: string;
}